see section 5.3 of the CONIKS paper
resolves #163