@Liamsi 2017-01-20
Note to myself: This PR still isn't really equivalent to what the java implementation did: We add 1000 users once but the inner "benchmark-loop" considers the newly added users in the first iteration only. 
Investigate if internally resetting 1000 hashes to `nil` works instead ([it should](https://github.com/coniks-sys/coniks-go/blob/0bedfc6e83b710377aee19011a90f48e57c1584d/merkletree/node.go#L78-L96)).