@c633 2017-03-06
> We add 1000 users once but the inner "benchmark-loop" considers the newly added users in the first iteration only.

I tried with this:
```
func benchPADUpdate(b *testing.B, entries uint64) {
	[...]
	// build the tree once:
	pad.Update(nil)

	// Insert 1000 additional entries (as described in section 5.3):
	var i uint64
	for i = 0; i < 1000; i++ {
		key := keyPrefix + string(i+entries)
		value := append(valuePrefix, byte(i+entries))
		if err := pad.Set(key, value); err != nil {
			b.Fatal(err)
		}
	}
	// clone current PAD's state:
	orgTree := pad.tree.Clone()
	b.ResetTimer()

	// now benchmark re-hashing the tree:
	for i := 0; i < b.N; i++ {
		b.StopTimer()
		pad.tree = orgTree.Clone()
		b.StartTimer()
		pad.Update(nil)
	}
}
```

There is the result of both versions:
```
# before
$ go test -bench BenchmarkPADUpdate100K
BenchmarkPADUpdate100K-4   	      10	 199899731 ns/op
PASS
ok  	github.com/coniks-sys/coniks-go/merkletree	215.545s

# after
$ go test -bench BenchmarkPADUpdate100K
BenchmarkPADUpdate100K-4   	      30	  67303366 ns/op
PASS
ok  	github.com/coniks-sys/coniks-go/merkletree	313.979s
```

What do you think @Liamsi ?