@masomel 2017-03-07
@c633 I like your solution!