@Liamsi 2017-03-07
I agree! Elegant and simple solution.